Fixing missing loading indicator when opening deployment/new
This happens when you open the new deployment modal, there is "Target
release" heading showing, but no options to select, and it looks like
the page has frozen.

Ensure the entire "Target release" panel is not shown in this case,
leaving the loading bar to show there is some activity going on in the
background.